RMI kann doch komplexe Objekte senden und Omni in omni umbenannt wegen Konvention

+++ src/main/java/server/users/Rollenverwaltung.java
@@ -102,7 +102,7 @@ public class Rollenverwaltung implements IRollenverwaltung {
         return mahnungen;
     }
 
-    public int getIdCounter() {
+    public long getIdCounter() {
         return IdCounter;
     }
